# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API APP",
        "description": "This is API for Nail App",
        "contact": {
            "email": "your-email@domain.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/voucher": {
            "get": {
                "tags": [
                    "voucher"
                ],
                "summary": "get all voucher",
                "operationId": "voucher_get",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "voucher"
                ],
                "summary": "Create voucher",
                "operationId": "voucher_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "voucher_name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "interger"
                                    },
                                    "quantity": {
                                        "type": "interger"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "voucher_name": "ILOVEYOU",
                                    "price": "5000",
                                    "quantity": "500"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/voucher-to-user": {
            "get": {
                "tags": [
                    "voucher"
                ],
                "summary": "get all voucher to user application",
                "operationId": "voucher_get_to_user_app",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/voucher/{id}": {
            "get": {
                "tags": [
                    "voucher"
                ],
                "summary": "get voucher by id",
                "operationId": "voucher_get_by_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "voucher"
                ],
                "summary": "Update voucher",
                "operationId": "voucher_put",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "voucher_name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "interger"
                                    },
                                    "quantity": {
                                        "type": "interger"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "voucher_name": "ILOVEYOU",
                                    "price": "5000",
                                    "quantity": "500"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "voucher"
                ],
                "summary": "Delete voucher ",
                "operationId": "voucher_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/total-by-voucher": {
            "post": {
                "tags": [
                    "voucher"
                ],
                "summary": "Total apply voucher",
                "operationId": "total_by_voucher",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "voucher_name": {
                                        "type": "string"
                                    },
                                    "total": {
                                        "type": "interger"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "voucher_name": "ILOVEYOU",
                                    "total": "500000"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "Bearer Token.",
                "in": "header",
                "scheme": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}